Temp undo VoiceOver fix for list-style-type: none
The VoiceOver fix for `list-style-type: none` was wrongly applied to the inline list item class. However, the error exposed a flaw in this fix. If inline-block is applied to the list item (as in the inline list), unwanted spacing is added. We might be able to visually hide the pseudo content, but this will require additional testing to verify it continues to work with VoiceOver.